+++ src/main/kotlin/com/macoder/styling/authentication/persistence/StylistRepository.kt
@@ -4,7 +4,7 @@ import com.macoder.styling.common.entity.Stylist
 import org.springframework.data.jpa.repository.JpaRepository
 
 
-interface StylistRepository: JpaRepository<Stylist, Int> {
+interface StylistRepository: JpaRepository<Stylist, Long> {
 
 
 }
